+++ CHANGELOG.md
@@ -21,6 +21,9 @@
 - Introduced `ysai` live template for Spring `@Autowired` declaration [#1384](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1384)
 - Introduced `ysaqi` live template for Spring `@Autowired` + `@Qualifier` declaration [#1385](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1385)
 
+### `Cockpit NG` enhancements
+- Re-calculate Cockpit NG Meta Model on some file operations [#1399](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1399)
+
 ### Fixes
 - Comply with _Companion object in extensions_ Plugin DevKit inspection [#1386](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1386)
 - Comply with _Cancellation exception handled incorrectly_ Plugin DevKit inspection [#1388](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1388)